Finished the Redux Auth And starteed the api calls for posts

+++ client/src/actions/auth.js
@@ -5,11 +5,14 @@ import {
   REGISTER_SUCCESS,
   GET_ERRORS,
   SET_LOGGED_USER,
+  GET_LOGGED_USER,
   CLEAR_CURRENT_USER,
-  DELETE_CURRENT_USER
+  DELETE_CURRENT_USER,
+  EDIT_CURRENT_USER,
+  EDITING_CURRENT_USER
 } from "./types";
 
-export const authRegisterUser = (user, history) => dispatch => {
+export const authRegisterUser = user => dispatch => {
   axios
     .post("/api/users/register", user)
     .then(res =>
@@ -47,12 +50,22 @@ export const authLoginUser = userData => dispatch => {
       })
     );
 };
+
 // Set logged in user
+
 export const setLoggedUser = decoded => {
   return {
     type: SET_LOGGED_USER,
     payload: decoded
   };
+};
+export const getCurrentUser = id => dispatch => {
+  axios.get(`/api/users/${id}`).then(res => {
+    dispatch({
+      type: GET_LOGGED_USER,
+      payload: res.data
+    });
+  });
 };
 // Clear profile
 export const clearCurrentUser = () => {
@@ -62,12 +75,33 @@ export const clearCurrentUser = () => {
 };
 // delete current user
 export const authDeleteUser = id => dispatch => {
-  axios.delete(`/api/users/${id}`).then(res => {
-    dispatch({
-      type: DELETE_CURRENT_USER,
-      payload: id
+  if (window.confirm("Are you sure?")) {
+    axios.delete(`/api/users/${id}`).then(res => {
+      dispatch({
+        type: DELETE_CURRENT_USER,
+        payload: id
+      });
     });
+  }
+};
+export const authEditUser = userData => dispatch => {
+  dispatch({
+    type: EDITING_CURRENT_USER
   });
+  axios
+    .put(`/api/users/edit/${userData.id}`, userData)
+    .then(res => {
+      dispatch({
+        type: EDIT_CURRENT_USER,
+        payload: {
+          user: res.data.result,
+          success: res.data.success
+        }
+      });
+    })
+    .catch(err => {
+      console.log(err);
+    });
 };
 
 // Log user out
